FEAT : 계정 삭제 로직에 비밀번호를 통해 검증 하도록 추가 #188

+++ src/main/java/com/todoay/api/domain/auth/service/AuthServiceImpl.java
@@ -79,10 +79,14 @@ public class AuthServiceImpl implements AuthService {
 
     @Transactional
     @Override
-    public void deleteAuth() {
+    public void deleteAuth(String password) {
         String email = jwtProvider.getLoginId();
         Auth auth =authRepository.findByEmail(email)
                 .orElseThrow(EmailNotFoundException::new);
+
+        if (!encoder.matches(password, auth.getPassword())) {
+            throw new PasswordNotMatchedException();
+        }
         auth.deleteAuth();
     }
